+++ src/index.ts
@@ -1,5 +1,3 @@
-/// <reference types="typescript" />
-
 export *                                                            from "./Styling";
 export {useComposedValues, useComposedStyle}                        from "./Hooks";
 export {default as CascadingValuesContext, CascadingValuesProvider} from "./CascadingValuesContext";